Fix flaky GCS ADC fallback test by isolating in own JVM
